+++ app/src/main/java/com/example/jetweather/views/currenthourweather/CurrentUvIndexCard.kt
@@ -31,7 +31,7 @@ fun CurrentUvIndexCard(viewModel: CurrentHourWeatherViewModel, modifier: Modifie
                 text = "UV Index",
                 iconId = R.drawable.cloud,
             )
-            Text(text = currentHourWeather.uvIndex.getUvIndexStatus())
+            Text(text = currentHourWeather.uvIndex.toInt().getUvIndexStatus())
         }
     }
 }
