+++ app/src/main/java/com/example/jetweather/weatherdata/CurrentHourWeatherData.kt
@@ -3,5 +3,5 @@ package com.example.jetweather.weatherdata
 data class CurrentHourWeatherData(
     val cloudCover: Int = 0,
     val visibility: Int = 0,
-    val uvIndex: Int = 0,
+    val uvIndex: Float = 0f,
 )
